Arquivo de configuração

+++ config.sh
@@ -1,12 +1,13 @@
 #! /bin/bash 
 
-rm fileOld.txt
-mv file.txt fileOld.txt 
 if (git diff-index --quiet --cached HEAD) ; then
+	if [-e fileOld.txt]; then
+		rm fileOld.txt
+	fi
+	mv file.txt fileOld.txt
 	ls -1 >> file.txt
 	diff fileOld.txt file.txt >> dif.txt 
 fi
-
 
 #while read line 
 #do
